+++ examples/symfony-app/.php-cs-fixer.dist.php
@@ -1,0 +1,6 @@
+<?php
+
+return (new PhpCsFixer\Config())
+    ->setRules([
+        '@Symfony' => true,
+    ]);