+++ examples/symfony-app/src/Controller/StatusController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Controller;
+
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
+
+class StatusController
+{
+    #[Route('/status', name: 'status')]
+    public function __invoke(): Response
+    {
+        return new Response('ok');
+    }
+}